`project` –> `git`
