+++ packages/cli/test/fixtures/unit/commands/git/connect/bad-remote-url/.vercel/project.json
