+++ packages/cli/test/fixtures/unit/commands/git/connect/existing-connection/.vercel/project.json
