+++ packages/cli/test/fixtures/unit/commands/git/connect/existing-connection/git/description
