+++ packages/cli/test/fixtures/unit/commands/git/connect/invalid-repo/.gitignore
