+++ packages/cli/test/fixtures/unit/commands/git/connect/new-connection/.gitignore
