+++ packages/cli/test/fixtures/unit/commands/git/connect/no-git-config/.vercel/project.json
