+++ packages/cli/test/fixtures/unit/commands/git/connect/same-repo-connection/.vercel/project.json
